network tests: fix FIFO behaviour, properly test partial reads
On the one hand, Afina may or may not close its write FIFO after sending
a reply. On the other hand, FIFO reads are atomic up to PIPE_BUF which is
>=4K and we "know" that our replies are way smaller. Therefore, just call
read() with a big enough buffer and let the kernel handle everything.

NB: this requires clarification (maybe Afina SHOULD close and reopen its
pipe when it detects EOF, after all)

Also, after sending a partially written command verify that it was properly
understood (including the body).